final clean up (hopefully)
